Fix warnings about 64-bit pointer to int32 convertion

+++ src/etnaviv/viv.h
@@ -134,7 +134,7 @@ enum viv_status
 };
 
 /* Type for GPU physical address */
-typedef uint32_t viv_addr_t;
+typedef uintptr_t viv_addr_t;
 
 /* General process handle */
 typedef uint64_t viv_handle_t;
